+++ packages/media-core/src/providers/mediaUrlPreviewProvider.ts
@@ -1,21 +1,23 @@
-import { MediaApiConfig, UrlPreview } from '../';
-import { Observable } from 'rxjs/Observable';
-import { UrlPreviewProvider } from './urlPreviewProvider';
-import { Pool } from './util/pool';
-import { observableFromObservablePool } from './util/observableFromObservablePool';
+import {MediaApiConfig, UrlPreview} from '../';
+import {Observable} from 'rxjs/Observable';
+import {UrlPreviewProvider} from './urlPreviewProvider';
+import {Pool} from './util/pool';
+import {observableFromObservablePool} from './util/observableFromObservablePool';
 
 export type MediaUrlPreviewObservablePool = Pool<Observable<UrlPreview>>;
 
 export class MediaUrlPreviewProvider extends UrlPreviewProvider {
-  public static fromMediaApi(config: MediaApiConfig, url: string, clientId: string) {
-    return UrlPreviewProvider.fromMediaApi(config, url, clientId);
+  public static fromMediaApi(config: MediaApiConfig, url: string) {
+    return UrlPreviewProvider.fromMediaApi(config, url);
   }
 
-  public static fromPool(pool: MediaUrlPreviewObservablePool, config: MediaApiConfig, url: string, clientId: string) {
+  public static fromPool(pool: MediaUrlPreviewObservablePool,
+                         config: MediaApiConfig,
+                         url: string) {
     return {
       observable() {
         return observableFromObservablePool(pool, url, () => {
-          return MediaUrlPreviewProvider.fromMediaApi(config, url, clientId).observable();
+          return MediaUrlPreviewProvider.fromMediaApi(config, url).observable();
         });
       }
     };