+++ packages/media-core/src/services/urlPreviewService.ts
@@ -2,17 +2,16 @@ import createRequest from './util/createRequest';
 import {UrlPreview, MediaApiConfig} from '../';
 
 export interface UrlPreviewService {
-  getUrlPreview(url: string, clientId: string): Promise<UrlPreview>;
+  getUrlPreview(url: string): Promise<UrlPreview>;
 }
 
 export class MediaUrlPreviewService implements UrlPreviewService {
 
   constructor(private readonly config: MediaApiConfig) {}
 
-  getUrlPreview(url: string, clientId: string): Promise<UrlPreview> {
+  getUrlPreview(url: string): Promise<UrlPreview> {
     const request = createRequest({
-      config: this.config,
-      clientId: clientId
+      config: this.config
     });
     return request({
       url: '/link/preview',