+++ packages/task-decision/src/components/ResourcedItemList.tsx
@@ -86,7 +86,7 @@ const LoadingWrapper = styled.div`
 
 export default class ResourcedItemList extends PureComponent<Props,State> {
   public static defaultProps: Partial<Props> = {
-    appearance: 'elevated'
+    appearance: 'card'
   };
 
   private mounted: boolean;
@@ -231,7 +231,12 @@ export default class ResourcedItemList extends PureComponent<Props,State> {
           if (isDecision(item)) {
             return (
               <li key={objectKeyToString(objectKey)}>
-                <DecisionItem appearance={appearance} >
+                <DecisionItem
+                  appearance={appearance}
+                  participants={item.participants}
+                  creator={item.creator}
+                  lastUpdater={item.lastUpdater}
+                >
                   {renderDocument(contentToDocument(item.content), toRendererContext(objectKey))}
                 </DecisionItem>
               </li>
@@ -248,6 +253,9 @@ export default class ResourcedItemList extends PureComponent<Props,State> {
                   objectAri={objectKey.objectAri}
                   containerAri={objectKey.containerAri}
                   appearance={appearance}
+                  participants={item.participants}
+                  creator={item.creator}
+                  lastUpdater={item.lastUpdater}
                 >
                   {renderDocument(contentToDocument(item.content), toRendererContext(objectKey))}
                 </ResourcedTaskItem>