+++ packages/task-decision/src/types.ts
@@ -21,11 +21,12 @@ export interface BaseItem<S> extends ObjectKey {
 export interface ServiceDecision {
   containerAri: string;
   creationDate: string;
-  creatorId: string;
+  creator?: User;
+  lastUpdater?: User;
   lastUpdateDate: string;
   localId: string;
   objectAri: string;
-  participants: any[];
+  participants: User[];
   // Atlassian Document fragment (json string)
   rawContent: string;
   state: DecisionState;
@@ -56,9 +57,10 @@ export interface ServiceTaskResponse {
 
 export interface Decision extends BaseItem<DecisionState> {
   creationDate: Date;
-  creatorId: string;
+  creator?: User;
+  lastUpdater?: User;
   lastUpdateDate: Date;
-  participants: any[];
+  participants: User[];
   // Atlassian Document fragment
   content: any;
   status: DecisionStatus;
@@ -91,15 +93,26 @@ export interface Query {
   sortCriteria?: SortCriteria;
 }
 
+// Kept for compatibility
+export type Participant = User;
+
+export interface User {
+  id: string;
+  displayName: string;
+  nickname: string;
+  avatarUrl: string;
+}
+
 export interface ServiceTask {
   containerAri: string;
   creationDate: string;
-  creatorId: string;
+  creator?: User;
+  lastUpdater?: User;
   lastUpdateDate: string;
   localId: string;
   objectAri: string;
   parentLocalId: string;
-  participants: any[];
+  participants: User[];
   position: number;
   // Atlassian Document fragment (json string)
   rawContent: string;
@@ -109,10 +122,11 @@ export interface ServiceTask {
 
 export interface Task extends BaseItem<TaskState> {
   creationDate: Date;
-  creatorId: string;
+  creator?: User;
+  lastUpdater?: User;
   lastUpdateDate: Date;
   parentLocalId: string;
-  participants: any[];
+  participants: User[];
   position: number;
   // Atlassian Document fragment
   content: any;
@@ -179,4 +193,4 @@ export interface OnUpdate<T> {
   (allDecisions: T[], newDecisions: T[]): void;
 }
 
-export type Appearance = 'elevated' | 'flat';
+export type Appearance = 'inline' | 'card';