+++ packages/xflow/src/common/notifyUsersAccessGranted.js
@@ -1,24 +1,9 @@
 import 'es6-promise/auto';
 import 'whatwg-fetch';
 
-import { getCurrentUsername, queryUsername, getInstanceName } from './tenantContext';
+import { getCurrentUsername, queryUsername, getInstanceName, getAvatarUrl } from './../common/tenantContext';
 
 export const NOTIFY_ENDPOINT_EAST = 'https://xflow.us-east-1.prod.public.atl-paas.net/accessgranted';
-const DEFAULT_AVATAR_URL = 'https://i2.wp.com/avatar-cdn.atlassian.com/default/96?ssl=1';
-const AVATAR_REGEXP = /^https:\/\/avatar-cdn.atlassian.com\/[A-Za-z0-9]+/;
-
-function getAvatarUrl({ avatarUrls }) {
-  // Find the largest size key
-  const key = Object.keys(avatarUrls || {}).pop();
-
-  if (!key) {
-    return DEFAULT_AVATAR_URL;
-  }
-
-  const baseUrl = (avatarUrls[key].match(AVATAR_REGEXP) || [])[0];
-  const url = baseUrl ? `${baseUrl}?s=128` : avatarUrls[key];
-  return url;
-}
 
 function getAtlassianAccountId({ attributes: { attributes } }) {
   if (!attributes) return '';
@@ -27,14 +12,14 @@ function getAtlassianAccountId({ attributes: { attributes } }) {
   return openIdAttr ? openIdAttr.values[0] : '';
 }
 
-async function notifyUsers(endpoint, instance, grantedAccessBy, grantedAccessTo) {
+async function notifyUsers(endpoint, instance, grantedAccessBy, grantedAccessTo, productKey) {
   const response = await fetch(endpoint, {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
     },
     body: JSON.stringify({
-      product: 'confluence',
+      product: productKey,
       instance,
       grantedAccessBy,
       grantedAccessTo,
@@ -50,7 +35,7 @@ async function notifyUsers(endpoint, instance, grantedAccessBy, grantedAccessTo)
   return await response.json();
 }
 
-export default async (users) => {
+export default async (users, productKey) => {
   if (users.length === 0) {
     return {
       status: 'SENT',
@@ -73,5 +58,11 @@ export default async (users) => {
     atlassianAccountId: getAtlassianAccountId(user),
   }));
 
-  return await notifyUsers(NOTIFY_ENDPOINT_EAST, instance, grantedAccessBy, grantedAccessTo);
+  return await notifyUsers(
+    NOTIFY_ENDPOINT_EAST,
+    instance,
+    grantedAccessBy,
+    grantedAccessTo,
+    productKey
+  );
 };